1631.Path_With_Minimum_Effort/1631. Path With Minimum Effort_test.go
package main

import "testing"

func Test_minimumEffortPath(t *testing.T) {
	type args struct {
		heights [][]int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			name: "Example1",
			args: args{heights: [][]int{{1, 2, 2}, {3, 8, 2}, {5, 3, 5}}},
			want: 2,
		},
		{
			name: "Example2",
			args: args{heights: [][]int{{1, 2, 3}, {3, 8, 4}, {5, 3, 5}}},
			want: 1,
		},
		{
			name: "Example3",
			args: args{heights: [][]int{{1, 2, 1, 1, 1}, {1, 2, 1, 2, 1}, {1, 2, 1, 2, 1}, {1, 2, 1, 2, 1}, {1, 1, 1, 2, 1}}},
			want: 0,
		},
		{
			name: "Example4",
			args: args{heights: [][]int{{4, 3, 4, 10, 5, 5, 9, 2}, {10, 8, 2, 10, 9, 7, 5, 6}, {5, 8, 10, 10, 10, 7, 4, 2}, {5, 1, 3, 1, 1, 3, 1, 9}, {6, 4, 10, 6, 10, 9, 4, 6}}},
			want: 5,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := minimumEffortPath(tt.args.heights); got != tt.want {
				t.Errorf("minimumEffortPath() = %v, want %v", got, tt.want)
			}
		})
	}
}
